+++ app/src/main/java/com/example/image_search/Adapter/FavoriteAdapter.kt
@@ -1,0 +1,4 @@
+package com.example.image_search.Adapter
+
+class FavoriteAdapter {
+}
