+++ app/src/main/java/com/example/image_search/Adapter/SearchAdapter.kt
@@ -1,0 +1,4 @@
+package com.example.image_search.Adapter
+
+class SearchAdapter {
+}
