+++ app/src/main/java/com/example/image_search/Data/SearchData.kt
@@ -1,4 +1,4 @@
-package com.example.image_search
+package com.example.image_search.Data
 
 import com.google.gson.annotations.SerializedName
 import java.util.Date
