+++ app/src/main/java/com/example/image_search/SearchFragment.kt
@@ -6,26 +6,26 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 
-private const val ARG_PARAM1 = "param1"
-private const val ARG_PARAM2 = "param2"
-
 class SearchFragment : Fragment() {
     private var param1: String? = null
     private var param2: String? = null
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        arguments?.let {
-            param1 = it.getString(ARG_PARAM1)
-            param2 = it.getString(ARG_PARAM2)
-        }
-    }
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_search_, container, false)
+        return inflater.inflate(R.layout.fragment_search, container, false)
+    }
+
+    private fun setImage(image: String): HashMap<String,String> {
+        val key = BuildConfig.KAKAO_REST_API
+        return hashMapOf(
+            "serviceKey" to key,
+            "returnType" to "json",
+            "numOfRows" to "80",
+            "pageNo" to "1",
+            "ver" to "1.0"
+        )
     }
 }
