--- app/src/main/res/layout/favorite_page.xml
@@ -1,41 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <ImageView
-        android:id="@+id/save_image_view"
-        android:layout_width="150dp"
-        android:layout_height="150dp"
-        android:src="@color/black"
-        android:layout_margin="20dp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        />
-
-    <TextView
-        android:id="@+id/title"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="[Image]네이버 블로그"
-        android:textStyle="bold"
-        android:textSize="12sp"
-        android:layout_marginTop="10dp"
-        app:layout_constraintTop_toBottomOf="@id/save_image_view"
-        app:layout_constraintStart_toStartOf="@id/save_image_view"
-        app:layout_constraintEnd_toEndOf="@id/save_image_view"
-        />
-
-    <TextView
-        android:id="@+id/time_text"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="2023-09-21 20:00:00"
-        app:layout_constraintStart_toStartOf="@id/title"
-        app:layout_constraintEnd_toEndOf="@id/title"
-        app:layout_constraintTop_toBottomOf="@id/title"
-        />
-
-
-</androidx.constraintlayout.widget.ConstraintLayout>
