+++ app/src/main/res/layout/fragment_search.xml
@@ -22,10 +22,11 @@
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"
             android:layout_width="240dp"
-            android:layout_height="50dp"
+            android:layout_height="45dp"
             android:layout_marginStart="20dp"
             android:background="@drawable/search_edit_text"
             android:hint="검색어를 입력해주세요."
+            android:textSize="15sp"
             android:padding="13dp" />
 
         <androidx.appcompat.widget.AppCompatButton
